+++ src/panoramic/cli/paths.py
@@ -69,6 +69,14 @@ class Paths:
         return package / SystemDirectory.FIELDS.value
 
     @staticmethod
+    def transforms_dir():
+        return Path.cwd() / SystemDirectory.TRANSFORMS.value
+
+    @staticmethod
+    def transforms_compiled_dir():
+        return Path.cwd() / SystemDirectory.TRANSFORMS.value / '.compiled'
+
+    @staticmethod
     def dataset_schema_file() -> Path:
         with importlib_resources.path(panoramic.cli.schemas, PresetFileName.DATASET_SCHEMA.value) as path:
             return path
@@ -113,6 +121,8 @@ class FileExtension(Enum):
 
     MODEL_YAML = '.model.yaml'
     FIELD_YAML = '.field.yaml'
+    TRANSFORM_YAML = '.transform.yaml'
+    COMPILED_TRANSFORM_SQL = '.transform.sql'
 
 
 class PresetFileName(Enum):
@@ -140,5 +150,6 @@ class PresetFileName(Enum):
 
 class SystemDirectory(Enum):
     SCANNED = 'scanned'
+    TRANSFORMS = 'transforms'
     FIELDS = 'fields'
     DBT = '.dbt'